E-mail filter for publish comments
In order to have more precise control of when
e-mail notifications are sent, a filter expression
can now be configured in site configuration.

An example filter configuration for Qt sanity
review bot:

[review]
	filter = SRVW < 0 || reviewer != qt_sanity_bot

If filter evaluates true the e-mail will be sent

=============
this should be rewritten to make use of gerrit's submit rule evaluator,
which has all the necessary facts about a change available.
=============

Task-number: QTQAINFRA-220
Originally-by: Tomas Ljunggren <tomas.ljunggren@tieto.com>
Change-Id: Ia23301f49397f017684ada7f7653781fd8770108
Reviewed-by: Ismo Haataja <ismo.haataja@digia.com>