Trim username in authentication request
Remove leading and trailing whitespace in username when creating a request
to avoid incorrect login. Valid username and extra whitespace with valid
password causes creation of new Gerrit user that logs in. This is unwanted
behavior and because whitespace is invalid character to use in username it
can be safely removed.

Task-number: QTQAINFRA-456
Change-Id: I0315285a2cc5b349515ba57c0d96167044238996
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Ismo Haataja <ismo.haataja@digia.com>