Fix gerrit staging-ls performance.
When searching open changes, full commit log was traversed and each change
was queried from database by the change-id found. Because we always list
the staging or build branch and know the destination branch, we can stop the
search when hitting the destination head. No need to go way back in history,
as over time this gets slower and slower.

Task-number: QTQAINFRA-568
Change-Id: I19e142f65e0996644dcbd50a08f61bbeb886dbf0
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Janne Anttila <janne.anttila@digia.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>