Don't create empty cherry-picks
If the tree after a cherry-pick is identical to the new parent's tree
the newly picked commit is empty.  This can happen if the commit is
being cherry-picked on top of another version of itself, or was
already included in the target branch.

Change-Id: I7e3fe9c02ca0b5182635a6e202b1d38e4dc39405
(cherry picked from commit fd889d9db648e6f027ee5d2e295041cb1bc3591a)
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>