[DROP] One page review.
A complete change can be reviewed from single page. The new screen is
launched from standard 'All Side-by-Side' and 'All Unified' links on the
change screen. One configuration change is needed in 'gerrit.config' to
make this happen:

[gerrit]
	onepagereview = true

If this value is not set (or is false) the default Gerrit way kicks in
and every file in change is opened in separate tab.

==============
this change is obsolete with PG, due to its ability to inline all diffs
directly on the change screen.

impl for CS2/CodeMirror: https://gerrit-review.googlesource.com/75968
it has shortcomings, but i'd ignore the performance issue (the
implementation here can block for minutes, too).

first upstreaming attempt: https://gerrit-review.googlesource.com/59320
==============

Originally-by: Mika Hamalainen <mika.hamalainen@accenture.com>
Change-Id: Ia075fdec58a86dec2534381e00fafb75e48d3d78
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Ismo Haataja <ismo.haataja@digia.com>