One page review.
A complete change can be reviewed from single page. The new screen is
launched from standard 'All Side-by-Side' and 'All Unified' links on the
change screen. One configuration change is needed in 'gerrit.config' to
make this happen:

[gerrit]
	onepagereview = true

If this value is not set (or is false) the default Gerrit way kicks in
and every file in change is opened in separate tab.

Originally-by: Mika Hamalainen <mika.hamalainen@accenture.com>
Change-Id: Ia075fdec58a86dec2534381e00fafb75e48d3d78
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Ismo Haataja <ismo.haataja@digia.com>