Commit message footer options for cherry-pick mode
When using cherry pick submit type, there is now additional options
which define which footers are added to the commit message. Also,
added footers can be limited to reviewers who gave the maximum
approval score. These options are visible only for actual projects,
not for All-Projects.

==============
https://bugs.chromium.org/p/gerrit/issues/detail?id=655

> These options are visible only for actual projects, not for
> All-Projects.
>
that's because this implementation pre-dates upstream's ability to have
projects inherit config settings. this feature should be obviously made
use of now.

> Also, added footers can be limited to reviewers who gave the maximum
> approval score.
>
that turned out to be a completely useless feature (replaced by the
sanity bot email suppression patch). please remove it.
==============

Bug: issue 655
Originally-by: Mika Hamalainen <mika.hamalainen@accenture.com>
Change-Id: Ibfb7193edde7f88607683adf061e8603cb382d06
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Ismo Haataja <ismo.haataja@digia.com>