Hide sanity review radio buttons
Sanity review category box now starts in hidden mode. All review
categories are now expandable boxes.

==============
this is hard to apply to both CS2 and PG, because of the new horizontal
arrangement of the score radios.
it would be possible to replace the radio groups with "..." buttons, but
that does a lousy job at "hiding" them, because the visual size is
essentially unaffected.
a more thorough measure would be putting a "more ..." button below the
visible-by-default labels, so the hidden labels would be actually
invisible by default.

the current implementation of this feature is an abomination, with
the label name being hard-coded. instead, the label configuration from
one of the previous patches should be extended by another column, so
you'd have one for the cherry-pick footers and one for default
visibility.
https://bugreports.qt.io/browse/QTQAINFRA-841
==============

Task-number: QTQAINFRA-355
Originally-by: Ola Bjuremo <ola.bjuremo@tieto.com>
Change-Id: I28ba61d00aec52da95d2ae6f890d18ed13fce2ac
Reviewed-by: Ismo Haataja <ismo.haataja@digia.com>