Fixed gerrit email verification link requiring insecure login
When registering a new email address with gerrit, it sends an email
with a verification link. This link now uses 'https:' schema instead
of 'http:'.

===============
setting gerrit.canonicalWebUrl (which we do now; dunno about when the
patch was created, but given that we always had a reverse proxy, it
seems unlikely that we didn't) is supposed to be sufficient according to
https://bugs.chromium.org/p/gerrit/issues/detail?id=1428
===============

Task-number: QTQAINFRA-167
Originally-by: Peter Heubeck <peter.heubeck@tieto.com>
Change-Id: I0e2e5a88171d6d6519149f5a2c2248810436c156
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Ismo Haataja <ismo.haataja@digia.com>