Hide the Rebase button
Permanently hide the Rebase button so rebasing a change is never
possible from the web UI. Set following value in 'gerrit.config'
to make this happen:

[gerrit]
    hiderebase = true

If this value is not set (or is false) the default Gerrit way is on.
Then the change owner and submitters can always rebase changes in the
web UI (even without having the Rebase access right assigned).

Task-number: QTQAINFRA-848
Change-Id: I6e26f6ab5e88d0446d6bcb5081e74976f16cc5d0
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Ismo Haataja <ismo.haataja@digia.com>