Remove incorrect error message after reviewing a change
If the state has changed to one of CI states (STAGING, STAGED or INTEGRATING)
during review, error message is shown to the user. This incorrectly happened also
when commenting a change already in CI state. Now the error message is shown only
when review was initiated from non CI state but the state has changed to one of
CI states meanwhile, before saving review.

Task-number: QTQAINFRA-858
Change-Id: I3be7b56c6e3d80f9b4c3f8a0a7a97a7757da4ac3
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>
Reviewed-by: Ismo Haataja <ismo.haataja@digia.com>