In dark mode, use a dark header bar
The blue is hard on the eyes in any case, with the green logo of ours,
it becomes horrible. Use some grey that is already in use for the
side-bar and everything looks better. I don't think there is an easy way
to override this in the plugin.

Fixes: QTQAINFRA-2997
Change-Id: I39664697f6c25c9bbd1fdce9eb9d21f6654940fe
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Jukka Jokiniva <jukka.jokiniva@qt.io>
(cherry picked from commit 6bfeeb4b88c87b8318d772113cf410a290f2e630)
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>