In dark mode, use a dark header bar
The blue is hard on the eyes in any case, with the green logo of ours,
it becomes horrible. Use some grey that is already in use for the
side-bar and everything looks better. I don't think there is an easy way
to override this in the plugin.

Fixes: QTQAINFRA-2997
Change-Id: I39664697f6c25c9bbd1fdce9eb9d21f6654940fe
(cherry picked from commit 46edaadfb6c826cf74a1bc381b2c06b851cf5cc0)
Reviewed-by: Jukka Jokiniva <jukka.jokiniva@qt.io>