Workaround for broken merge commit meta history
This patch is a temporary workaround for displaying
merge commits in Gerrit. The workflow plugin
worked around a previous limitation of gerrit when
attempting to merge "merge commits"; This resulted
in an invalid patchset in the meta history which
references the same sha as the final user-uploaded
patchset. Changes in gerrit upstream that provide
new functionality for fetching parents of a change
broke when encountering this duplicate/virtual
patchset.

Task-number: QTQAINFRA-7338
Change-Id: I995072b6354cf95dc1492e8342a744eb6e752b13
Reviewed-by: Jukka Jokiniva <jukka.jokiniva@qt.io>